+++ src/menu.h
@@ -21,8 +21,8 @@ class Menu {
         std::string logo{};
         int width, height;
         struct {
-            uint8_t x;
-            uint8_t y;
+            uint16_t x;
+            uint16_t y;
         } logo_offset;
         int logo_width, logo_height;
         WINDOW* title;
@@ -30,12 +30,12 @@ class Menu {
 
         // functions for drawing
         void parse_logo(std::string& buffer);
-        bool is_frame(uint8_t row, uint8_t col);
+        bool is_frame(uint16_t row, uint16_t col);
         void fill_screen();
         void draw_logo();
         void draw_frame();
         void draw_content();
-        void set_parameters(); 
+        void set_parameters();
         //   ^ sets parameters for drawing based i.e. line width,
         //     element spacing based on window size
